♻️ F01c: HOME > FeaturedCollection > Create <FeaturedContent/>
