+++ src/components/reusable-ui/Title.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { theme } from '../../theme';
+import { theme } from '../../theme'
 
 export default function Title({ value }) {
   return (
